--- after.sh
@@ -1,5 +1,0 @@
-#!/bin/sh
-
-# If you would like to do some extra provisioning you may
-# add any commands you wish to this file and they will
-# be run after the Homestead machine is provisioned.
