+++ app/Attachment.php
@@ -3,7 +3,6 @@
 namespace App;
 
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\SoftDeletes;
 use Storage;
 
 /**
@@ -32,10 +31,6 @@ use Storage;
  */
 class Attachment extends Model
 {
-    use SoftDeletes;
-
-    protected $dates = ['deleted_at'];
-
     /**
      * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
      */
