+++ app/Http/Controllers/UserController.php
@@ -5,8 +5,6 @@ namespace App\Http\Controllers;
 use App\Attachment;
 use App\MobilityStatus;
 use App\Role;
-use App\Semester;
-use App\UniversityBranch;
 use Illuminate\Http\Request;
 use JavaScript;
 use App\User;
@@ -48,9 +46,7 @@ class UserController extends Controller
                 'mobilities.universityBranch.country',
                 'bank_accounts'])->find($id),
             'attachments' => is_null($active_mobility) ? null : Attachment::where('mobility_id', $active_mobility->id)->get(),
-            'mobility_statuses' => MobilityStatus::all(),
-            'semesters' => Semester::all(),
-            'university_branches' => UniversityBranch::all()
+            'mobilityStatuses' => MobilityStatus::all()
         ]);
         return View::make('view.user', ['user_id' => $id]);
     }
