+++ composer.json
@@ -16,10 +16,9 @@
     "require-dev": {
         "filp/whoops": "~2.0",
         "fzaninotto/faker": "~1.4",
-        "laravel/homestead": "^7.0",
         "mockery/mockery": "~1.0",
-        "phpspec/phpspec": "~2.0",
-        "phpunit/phpunit": "~6.0"
+        "phpunit/phpunit": "~6.0",
+        "phpspec/phpspec": "~2.0"
     },
     "autoload": {
         "classmap": [
