+++ package.json
@@ -13,25 +13,24 @@
         "laraclean": "php artisan cache:clear && php artisan config:clear && php artisan route:clear && php artisan view:clear && php artisan clear-compiled"
     },
     "devDependencies": {
-        "axios": "^0.17.1",
+        "axios": "^0.16.2",
         "babel-helper-vue-jsx-merge-props": "^2.0.3",
         "babel-plugin-syntax-jsx": "^6.18.0",
         "babel-plugin-transform-vue-jsx": "^3.5.0",
         "babel-preset-env": "^1.6.1",
         "bootstrap-sass": "^3.3.7",
-        "cross-env": "^5.1.3",
+        "cross-env": "^5.0.1",
         "jquery": "^3.1.1",
-        "laravel-mix": "^1.7.2",
+        "laravel-mix": "^1.0",
         "lodash": "^4.17.4",
-        "vue": "^2.5.13"
+        "vue": "^2.1.10"
     },
     "dependencies": {
-        "element-ui": "^2.0.11",
+        "element-ui": "^2.0.8",
         "font-awesome": "^4.7.0",
         "iban": "0.0.8",
-        "vue-form-wizard": "^0.8.2",
-        "vue-moment": "^3.1.0",
-        "vue-tables-2": "^1.2.40"
+        "vue-form-wizard": "^0.8.1",
+        "vue-tables-2": "^1.0.6"
     },
     "name": "uri-dashboard",
     "version": "1.0.0",
