+++ resources/assets/js/vue/vue-view-user.js
@@ -3,13 +3,12 @@ if ($('#vue-view-user').length) {
     const VueViewUser = new Vue({
         el: '#vue-view-user',
         data: {
-            ready: true,
-            editMobility: false,
-            user: DataFromBackend.user,
-            attachments: DataFromBackend.attachments,
-            mobility_statuses: DataFromBackend.mobility_statuses,
-            semesters: DataFromBackend.semesters,
-            university_branches: DataFromBackend.university_branches,
+            ready: false,
+            activeTab: 'registry',
+            user: window.user,
+            attachments: window.attachments,
+            attachment_description: '',
+            mobility_statuses: window.mobilityStatuses,
             bankTableColumns: ['is_main', 'bank_name', 'iban', 'holder_name', 'holder_surname'],
             bankTableOptions: {
                 filterable: false,
@@ -34,6 +33,20 @@ if ($('#vue-view-user').length) {
             },
         },
         computed: {
+            attachment_additions: function () {
+                return {
+                    data: {
+                        mobility_id: this.user.mobilities[0].id,
+                        description: this.attachment_description
+                    },
+                    headers: {
+                        'X-CSRF-TOKEN': document.head.querySelector('meta[name="csrf-token"]').getAttribute('content')
+                    }
+                }
+            },
+            mobilityStatusActiveTab: function () {
+                return this.has_mobilities ? this.user.mobilities[0].mobility_status_id - 1 : 0;
+            },
             full_name: function () {
                 return !this.ready ? '...' : this.user.name + ' ' + this.user.middle_name + ' ' + this.user.surname;
             },
@@ -57,7 +70,82 @@ if ($('#vue-view-user').length) {
                         console.log(response);
                     });
                 this.user.active_bank_account_id = e.row.id;
+            },
+            uploadAttachment: function () {
+                this.$refs.attachmentUploader.submit();
+            },
+            onAttachmentUploaded: function (response, file, fileList) {
+                this.$message({
+                    type: response.status,
+                    message: response.message
+                });
+                this.attachments.push(response.file);
+            },
+            onAttachmentClicked: function (file, column, cell, event) {
+                switch(column.label) {
+                    case "Download":
+                        this.downloadAttachment(file);
+                        break;
+                    case "Elimina":
+                        this.removeAttachment(file);
+                        break;
+                    default:
+                        break;
+                }
+            },
+            downloadAttachment: function (file) {
+                axios.post('/file/retrieve/', {id: file.id})
+                    .then(response => {
+                        this.$message({
+                            type: response.data.status,
+                            message: response.data.message
+                        });
+                        window.open(response.data.url);
+                    })
+                    .catch(error => {
+                        console.log(error.response.data);
+                    });
+            },
+            removeAttachment: function (file) {
+                // TODO implement attachment removal
+            },
+            mobilityNextStep() {
+                this.$confirm('Portare la mobilità al prossimo stato?', 'Info', {
+                    confirmButtonText: 'OK',
+                    cancelButtonText: 'Annulla',
+                    type: 'info'
+                }).then(() => {
+                    axios.put('/edit/mobility', {
+                        id: this.user.mobilities[0].id,
+                        new_status_id: this.user.mobilities[0].mobility_status_id + 1
+                    })
+                        .then((response) => {
+                            this.user.mobilities[0].mobility_status_id++;
+                            this.$message({
+                                type: response.data.status,
+                                message: response.data.message
+                            });
+                        })
+                        .catch((response) => {
+                            console.log(response);
+                            this.$message({
+                                type: response.data.status,
+                                message: response.data.message
+                            });
+                        });
+                }).catch(() => {
+                    this.$message({
+                        type: 'info',
+                        message: 'Avanzamento annullato'
+                    });
+                });
+            },
+            navigateToTab(navigateFunction, index) {
+                // TODO direct navigation between statuses not implemented
             }
+        },
+        mounted: function () {
+            this.ready = true;
         }
     });
 }
