+++ resources/views/content/the_sidebar.blade.php
@@ -40,7 +40,10 @@
                 </li>
             @endif
             <li>
-                <a href="#"><i class="fa fa-fw fa-bank"></i> Altro...</a>
+                <a href="#"><i class="fa fa-fw fa-bank"></i> Page four</a>
+            </li>
+            <li>
+                <a href="#"><i class="fa fa-fw fa-dropbox"></i> Page 5</a>
             </li>
             <li>
                 <a href="#" onclick="event.preventDefault();
